--- starters/vue3-apollo-quasar/src/variables/counts.js
@@ -1,3 +1,0 @@
-import { makeVar } from '@apollo/client';
-
-export const counts = makeVar(0);
